requirements.txt
pandas
phidata
duckduckgo-search
fastapi
groq
yfinance
python-multipart
uvicorn